requirements.txt
# Основные зависимости для сайта РостФеррум
fastapi==0.104.1
uvicorn[standard]==0.24.0
jinja2==3.1.2
python-multipart==0.0.6
psycopg[binary]==3.2.10

# Для интеграции с Битрикс24 (используем httpx вместо aiohttp)
httpx==0.25.2
aiofiles==24.1.0

# Для работы с данными (только если нужен импорт каталога)
# pandas==2.1.3
# openpyxl==3.1.2

# Для работы с изображениями (обработка превью/конвертация)
Pillow==10.4.0
 
# Загрузка переменных окружения из .env
python-dotenv==1.0.1

# Прод-сервер (Gunicorn + UvicornWorker)
gunicorn==21.2.0